[4633] Move degree creation into separate file
g
